add HTTPError in SnabblePhoneAuth
