+++ Sources/Network/HTTPError.swift
@@ -7,19 +7,8 @@
 
 import Foundation
 
-public enum HTTPError: LocalizedError {
+public enum HTTPError: Error {
     case invalid(HTTPURLResponse)
     case unknown(URLResponse)
     case unexpected(Error)
-    
-    public var errorDescription: String? {
-        switch self {
-        case let .invalid(response):
-            return "Error: statusCode: \(response.httpStatusCode.rawValue)"
-        case let .unknown(response):
-            return "Error: unknown \(response)"
-        case .unexpected:
-            return "Error: unexpected should not happen"
-        }
-    }
 }
